Finish whole map parsing

+++ srcs/parsing_by_cross.c
@@ -6,17 +6,17 @@
 /*   By: vismaily <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/02 20:14:15 by vismaily          #+#    #+#             */
-/*   Updated: 2021/11/04 19:05:58 by vismaily         ###   ########.fr       */
+/*   Updated: 2021/11/13 20:33:43 by vismaily         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3D.h"
 
-static int parsing_bottom(size_t i, size_t j)
+static int	parsing_bottom(size_t i, size_t j)
 {
 	size_t	k;
 	size_t	len;
-	
+
 	k = i;
 	len = my_map.map_height;
 	if (i == len - 1 && my_map.map[i][j] != '1' && my_map.map[i][j] != ' ')
@@ -31,10 +31,10 @@ static int parsing_bottom(size_t i, size_t j)
 	return (0);
 }
 
-static int parsing_top(size_t i, size_t j)
+static int	parsing_top(size_t i, size_t j)
 {
-	size_t	k;
-	
+	int	k;
+
 	k = i;
 	if (i == 0 && (my_map.map[i][j] != '1' && my_map.map[i][j] != ' '))
 		return (errors(6));
@@ -48,11 +48,11 @@ static int parsing_top(size_t i, size_t j)
 	return (parsing_bottom(i, j));
 }
 
-static int parsing_rigth(size_t i, size_t j)
+static int	parsing_rigth(size_t i, size_t j)
 {
 	size_t	k;
 	size_t	len;
-	
+
 	k = j;
 	len = ft_strlen(my_map.map[i]);
 	if (j == len - 1 && my_map.map[i][j] != '1' && my_map.map[i][j] != ' ')
@@ -69,7 +69,7 @@ static int parsing_rigth(size_t i, size_t j)
 
 int	parsing_by_cross(size_t i, size_t j)
 {
-	size_t	k;
+	int	k;
 
 	k = j;
 	if (j == 0 && (my_map.map[i][j] != '1' && my_map.map[i][j] != ' '))
